split records tasks from convert + linting
