+++ pycytominer_transform/__init__.py
@@ -2,16 +2,12 @@
 __init__.py for pycytominer_convert
 """
 from .convert import (
-    build_path,
     concat_join_records,
     concat_record_group,
     convert,
-    filter_source_filepaths,
     gather_records,
     get_join_chunks,
-    get_source_filepaths,
     infer_record_group_common_schema,
-    infer_source_datatype,
     join_record_chunk,
     prepend_column_name,
     read_data,
@@ -25,4 +21,10 @@ from .exceptions import (
     SchemaException,
 )
 from .presets import config
+from .records import (
+    build_path,
+    filter_source_filepaths,
+    get_source_filepaths,
+    infer_source_datatype,
+)
 from .utils import column_sort
